+++ src/components/Input/SelectInput/index.ts
@@ -1,0 +1,2 @@
+export * from "./SelectInput.component";
+export type * from "./SelectInput.types";