+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `v0.0.1` SelectInput
+  - A customizable, searchable select input
+
+### Changed
+
+- `mergeDefaultProps`, `useMergedProps`
+  - Allow foreign props not contained in defaultProps in final object.
+  - Useful for polymorphic props
+
+- `FormFieldDisplay`
+  - display selected option's label based on field value
+
 ## 0.20.1 - 2024-07-23
 ### Fixed
 - `v0.0.4` CopyClipboardInfoCard